--- src/main/kotlin/com/innoq/event/blockchainktwebflux/BlocksHandler.kt
@@ -1,16 +1,0 @@
-package com.innoq.event.blockchainktwebflux
-
-import com.innoq.event.blockchainktwebflux.domain.BlockChain
-import org.springframework.web.reactive.function.server.ServerRequest
-import org.springframework.web.reactive.function.server.ServerResponse
-import org.springframework.web.reactive.function.server.ServerResponse.*
-import org.springframework.web.reactive.function.server.body
-import reactor.core.publisher.Mono
-
-class BlocksHandler(val chain: BlockChain) {
-
-    fun showBlocks(request: ServerRequest): Mono<ServerResponse> {
-        return ok().body(Mono.just(chain))
-    }
-
-}
