--- src/main/kotlin/com/innoq/event/blockchainktwebflux/NodeInfoHandler.kt
@@ -1,13 +1,0 @@
-package com.innoq.event.blockchainktwebflux
-
-import com.innoq.event.blockchainktwebflux.domain.NodeInfo
-import org.springframework.web.reactive.function.server.ServerRequest
-import org.springframework.web.reactive.function.server.ServerResponse
-import org.springframework.web.reactive.function.server.body
-import reactor.core.publisher.Mono
-
-class NodeInfoHandler(val nodeInfo: NodeInfo) {
-
-    fun index(request: ServerRequest): Mono<ServerResponse>
-            = ServerResponse.ok().body(Mono.just(nodeInfo))
-}
